feat: Adds users attribute to mongodbatlas_projects plural data source (#3451)
* WIP

* Fixed acceptance test

* Added a comment

* Added changelog

* Added users attribute to projects data source

* Added changelog and fixed docs

* Factored out users schema

* Refactored schemas and fixed test and doc

* Minor code review suggestions and fixed doc

---------
