+++ .changelog/3451.txt
@@ -1,0 +1,3 @@
+```release-note:enhancement
+data-source/mongodbatlas_projects: Adds `users` attribute
+```
